Add icons to hamburger menu links
